combine endpoints
