+++ apps/web/package.json
@@ -85,7 +85,7 @@
     "rehype-slug": "^5.1.0",
     "remark-gfm": "^3.0.1",
     "shiki": "^0.14.1",
-    "sonner": "^0.5.0",
+    "sonner": "^1.4.41",
     "stripe": "^12.12.0",
     "swr": "^2.1.5",
     "topojson-client": "^3.1.0",
